requirements.txt
fastapi
uvicorn
firebase-admin
pydantic-settings
python-dotenv
aioredis==1.3.1
